# py/hmm_service/docs/api_specification.yaml
openapi: 3.0.3
info:
  title: HMM Microservice API
  description: |
    Hidden Markov Model inference service for the IMP trading system.
    
    This service provides real-time HMM state probability calculation and fusion weight computation
    for regime-aware signal fusion in trading applications.
    
    ## Features
    - State probability calculation from observation vectors
    - Fusion weight computation for signal combination
    - Model hot-reloading without service downtime
    - Comprehensive health monitoring and metrics
    - Structured error handling with detailed diagnostics
    
    ## Performance Targets
    - Inference latency: <20ms (p95)
    - Health check latency: <5ms
    - Concurrent requests: 100+ req/sec
    - Availability: 99.9%
    
    ## Authentication
    Currently no authentication is required. In production, consider implementing:
    - API key authentication
    - IP whitelisting
    - Rate limiting per client
  version: 1.0.0
  contact:
    name: IMP Development Team
    email: dev@imp-trading.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://hmm-service.imp-trading.com
    description: Production server
paths:
  /inference/state-probabilities:
    post:
      tags:
        - Inference
      summary: Calculate HMM state probabilities
      description: |
        Calculate Hidden Markov Model state probabilities for a given observation vector.
        
        The observation vector should contain exactly 3 values representing:
        - s_LDC: Lorentzian Distance Classifier signal
        - s_MR: Mean Reversion signal  
        - s_TSMOM: Time Series Momentum signal
        
        Returns normalized probabilities that sum to 1.0.
      operationId: calculateStateProbabilities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InferenceRequest'
            examples:
              basic_request:
                summary: Basic inference request
                value:
                  observations: [0.75, -0.32, 0.18]
                  timestamp: 1640995200
              with_request_id:
                summary: Request with tracking ID
                value:
                  observations: [0.45, 0.12, -0.67]
                  timestamp: 1640995260
                  request_id: "req-12345"
      responses:
        '200':
          description: State probabilities calculated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateProbabilitiesResponse'
              examples:
                successful_response:
                  summary: Successful state probability calculation
                  value:
                    state_probabilities: [0.15, 0.65, 0.20]
                    most_likely_state: 1
                    confidence: 0.65
                    timestamp: 1640995200
                    processing_time_ms: 12.5
                    request_id: "req-12345"
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /inference/fusion-weights:
    post:
      tags:
        - Inference
      summary: Calculate fusion weights
      description: |
        Calculate fusion weights for signal combination based on current market regime.
        
        This endpoint first calculates state probabilities, then computes corresponding
        fusion weights for each signal type (LDC, MR, TSMOM) based on the per-state
        weight matrices loaded from the HMM model.
        
        Weights are normalized and bounded within [-1.0, 1.0].
      operationId: calculateFusionWeights
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InferenceRequest'
            examples:
              basic_request:
                summary: Basic fusion weights request
                value:
                  observations: [0.75, -0.32, 0.18]
                  timestamp: 1640995200
      responses:
        '200':
          description: Fusion weights calculated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FusionWeightsResponse'
              examples:
                successful_response:
                  summary: Successful fusion weights calculation
                  value:
                    weights:
                      w_ldc: 0.45
                      w_mr: 0.30
                      w_tsmom: 0.25
                    state_probabilities: [0.15, 0.65, 0.20]
                    most_likely_state: 1
                    timestamp: 1640995200
                    processing_time_ms: 15.2
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /inference/predict:
    post:
      tags:
        - Inference
      summary: Complete HMM prediction
      description: |
        Get complete HMM prediction including both state probabilities and fusion weights.
        
        This is the most comprehensive endpoint that provides all inference results
        in a single request, including model version information and detailed metadata.
        
        Recommended for production use when you need both state probabilities and weights.
      operationId: predict
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InferenceRequest'
            examples:
              complete_request:
                summary: Complete prediction request
                value:
                  observations: [0.75, -0.32, 0.18]
                  timestamp: 1640995200
                  request_id: "trading-signal-001"
      responses:
        '200':
          description: Complete prediction generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionResponse'
              examples:
                successful_response:
                  summary: Complete prediction response
                  value:
                    state_probabilities: [0.15, 0.65, 0.20]
                    most_likely_state: 1
                    confidence: 0.65
                    fusion_weights:
                      w_ldc: 0.45
                      w_mr: 0.30
                      w_tsmom: 0.25
                    timestamp: 1640995200
                    processing_time_ms: 18.7
                    model_version: "v1.2.0"
                    request_id: "trading-signal-001"
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  schemas:
    InferenceRequest:
      type: object
      required:
        - observations
      properties:
        observations:
          type: array
          items:
            type: number
            format: float
          minItems: 3
          maxItems: 3
          description: "Observation vector [s_LDC, s_MR, s_TSMOM]"
          example: [0.75, -0.32, 0.18]
        timestamp:
          type: integer
          format: int64
          description: "Unix timestamp for the observation"
          example: 1640995200
        request_id:
          type: string
          description: "Optional request identifier for tracking"
          example: "req-12345"
    StateProbabilitiesResponse:
      type: object
      required:
        - state_probabilities
        - most_likely_state
        - confidence
        - timestamp
        - processing_time_ms
      properties:
        state_probabilities:
          type: array
          items:
            type: number
            format: float
            minimum: 0.0
            maximum: 1.0
          description: "Normalized probability distribution over HMM states"
          example: [0.15, 0.65, 0.20]
        most_likely_state:
          type: integer
          minimum: 0
          description: "Index of the most likely state"
          example: 1
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: "Confidence score (maximum probability)"
          example: 0.65
        timestamp:
          type: integer
          format: int64
          description: "Response timestamp"
          example: 1640995200
        processing_time_ms:
          type: number
          format: float
          minimum: 0.0
          description: "Processing time in milliseconds"
          example: 12.5
        request_id:
          type: string
          description: "Request identifier if provided"
          example: "req-12345"
    FusionWeightsResponse:
      type: object
      required:
        - weights
        - state_probabilities
        - most_likely_state
        - timestamp
        - processing_time_ms
      properties:
        weights:
          type: object
          properties:
            w_ldc:
              type: number
              format: float
              minimum: -1.0
              maximum: 1.0
              description: "Fusion weight for LDC signal"
            w_mr:
              type: number
              format: float
              minimum: -1.0
              maximum: 1.0
              description: "Fusion weight for Mean Reversion signal"
            w_tsmom:
              type: number
              format: float
              minimum: -1.0
              maximum: 1.0
              description: "Fusion weight for Time Series Momentum signal"
          required: [w_ldc, w_mr, w_tsmom]
          example:
            w_ldc: 0.45
            w_mr: 0.30
            w_tsmom: 0.25
        state_probabilities:
          type: array
          items:
            type: number
            format: float
            minimum: 0.0
            maximum: 1.0
          description: "State probabilities used for weight computation"
          example: [0.15, 0.65, 0.20]
        most_likely_state:
          type: integer
          minimum: 0
          description: "Index of the most likely state"
          example: 1
        timestamp:
          type: integer
          format: int64
          description: "Response timestamp"
          example: 1640995200
        processing_time_ms:
          type: number
          format: float
          minimum: 0.0
          description: "Processing time in milliseconds"
          example: 15.2
        request_id:
          type: string
          description: "Request identifier if provided"
          example: "req-12345"
    PredictionResponse:
      type: object
      required:
        - state_probabilities
        - most_likely_state
        - confidence
        - fusion_weights
        - timestamp
        - processing_time_ms
        - model_version
      properties:
        state_probabilities:
          type: array
          items:
            type: number
            format: float
            minimum: 0.0
            maximum: 1.0
          description: "Probability distribution over HMM states"
          example: [0.15, 0.65, 0.20]
        most_likely_state:
          type: integer
          minimum: 0
          description: "Index of the most likely state"
          example: 1
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: "Confidence score (maximum probability)"
          example: 0.65
        fusion_weights:
          type: object
          properties:
            w_ldc:
              type: number
              format: float
              minimum: -1.0
              maximum: 1.0
            w_mr:
              type: number
              format: float
              minimum: -1.0
              maximum: 1.0
            w_tsmom:
              type: number
              format: float
              minimum: -1.0
              maximum: 1.0
          required: [w_ldc, w_mr, w_tsmom]
          example:
            w_ldc: 0.45
            w_mr: 0.30
            w_tsmom: 0.25
        timestamp:
          type: integer
          format: int64
          description: "Response timestamp"
          example: 1640995200
        processing_time_ms:
          type: number
          format: float
          minimum: 0.0
          description: "Processing time in milliseconds"
          example: 18.7
        model_version:
          type: string
          description: "Version of the HMM model used"
          example: "v1.2.0"
        request_id:
          type: string
          description: "Request identifier if provided"
          example: "req-12345"
...
  responses:
    ValidationError:
      description: Validation error in request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_observations:
              summary: Invalid observation vector
              value:
                error: "VALIDATION_ERROR"
                error_code: "INVALID_OBSERVATIONS"
                message: "Observations must contain exactly 3 numeric values"
                timestamp: 1640995200
                details:
                  field: "observations"
                  provided_length: 2
                  required_length: 3
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internal_error:
              summary: Internal server error
              value:
                error: "INTERNAL_ERROR"
                error_code: "INTERNAL_500"
                message: "An internal error occurred"
                timestamp: 1640995200
    ServiceUnavailable:
      description: Service temporarily unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            service_unavailable:
              summary: Service overloaded
              value:
                error: "SERVICE_UNAVAILABLE"
                error_code: "SERVICE_OVERLOADED"
                message: "Service is temporarily overloaded, please retry"
                timestamp: 1640995200
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (future implementation)
...